moved the utility functions to a separate module
